Refactor: Move AuthProvider and update tsconfig
