+++ app/auth/login.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { View, Text, StyleSheet, TextInput, TouchableOpacity, Alert } from 'react-native';
 import { useRouter } from 'expo-router';
 import { Smartphone, Lock } from 'lucide-react-native';
-import { useAuth } from '@/hooks/useAuth';
+import { useAuth } from '../../providers/AuthProvider';
 
 export default function LoginScreen() {
   const router = useRouter();
